Added multiprocessing support
Added OpenMP dependencies in CMakeLists.txt, and modified the detection
loop in AutoBubStart3.cpp to run in parallel.  This makes the detection
stage far faster (using abub3hs on 7 sample runs previously timed out at
an hour, now it completes in ~20 minutes).

In order to write the data in event order (without modifying other
files), the PICO60Output object has to be deleted before the detect
loop, and each loop iteration creates and destroys its own OutputWriter.

In principle the training could be done in parallel too, however OpenMP
doesn't play well when raising errors within try...catch blocks, so if
the training fails (for e.g. if training data for cams 2 and 3 don't
exist, as in all data up to July 13, 2020) the program crashes.