Fixed frame0 being off by one
LocalizeOMatic had a line which read the previously identified trigger
frame into a cv::Mat object, but for some reason incremented the index
by 1.  This meant that the pixel coordinates for each bubble
corresponded to the frame after the one recorded in the output file.
e.g. if frame 50 were the trigger frame, the coordinates correspond to
the bubbles' position in frame 51, but frame0 still said 50.

Removed a line which incremented the trigger frame for cam2.  This was
likely a remnant from PICO-60, in which cam2 had focusing issues.  There
is still a lot of code which references cam2 separately from the other
three cameras.