Added mask directory AnalyzerUnit/L3Localizer
Previously, the mask images had to be stored in the same directory as
the abub3 executable, as the argument passed to cv2::imread was
"./cam_masks/cam?_mask.bmp".  For some reason this stopped working, and
the executable had to be run while in the directory.

To fix this, an additional required argument was added, the mask
directory, and AnalyzerUnit and L3Localizer classes were modified to
store the string.