Moved calculateEnetropyFrame to members of classes
When trying to parallelize AutoBub in the past, the results were
inconsistent in that running it twice would result in two different
files.  This was found to be due to the calculateEntropyFrame function,
which was being called from different places at the same time and
overwriting the image in use.  The function is only called from two
places inside classes, so it has been moved to be member functions of
those two classes so that each instance has its own copy of the
function.

AutoBubStart3 has been modified to process each camera in an event in
parallel.