Starting support for different storage conventions
AutoBub3 was written with the image storage convention used in PICO-60,
i.e. the images are stored in a subdirectory of the event folder with
file name "cam?_image?.png". Older experiments stored the images in the
event directory with the file name "cam?image?.bmp", where the latter
question mark is a space-padded right-adjusted integer with up to three
figures.

The main function now accepts an additional command line argument to
specify the data series, which determines the image storage convention
(both the image storage folder and file name convention).  The storage
convention is then passed to the Trainer class, which it uses to
retrieve the image file list.

Both the Trainer and AnalyzerUnit classes used a separate but identical
function to sort the image files by frame number.  This function was
hard-coded to use the PICO-60 convention, and std::sort prevents the use
of a member function of either Trainer or AnalyzerUnit to sort the list
of frames.  The frame sorter was changed to a class which uses an
overloaded call operator ( bool operator() ) to sort the image list by
frame.