Fixed output in edge cases with <3 cameras
The output files when there are fewer than three cameras sometimes has
lines for four cameras for an event (though since nothing has been
staged to be written, the data and camera number are random for the
lines corresponding to missing cameras).  This seems to have happened
when there is an error code written.

Also removed the hard-coded 30 frame offset between the internal image
frame counter and the output (PICO-60 began at frame 30, other chambers
begin at frame 0).  Now depends on the data-series parameter passed as a
cmd argument.