Removed code which prevented late bubble identification
A section of code in LocalizeOMatic prevented bubbles from being
identified or tracked if the trigger frame were within 5 frames of final
frame.  In PICO-60 and 40L data, this means any time the bubble was
first observed in a camera after frame 66 was ignored.  This also
prevented index errors, so an additional check is required after the
code which was removed.