Made SearchPattern public, reduced file restrictions
The SearchPattern member of Trainer was made public so that AnalyzerUnit
can use it to parse and sort images by camera.

File restrictions were reduced when checking whether an image is
malformed.  Both the number of files required in the image set and the
file size per image were reduced, so that smaller chambers (which
usually take fewer, smaller images) can use AutoBub.