Fixed output when trainers fail to train
OpenMP apparently does not work well with try-catch statements, at least
in the way they were being used by the Trainer class.  If a Trainer
found no images (such as in early PICO data) it would throw with the
intent of the main program catching the exception.  OpenMP makes this
difficult, so instead the previously-unused StatusCode attribute of the
Trainer class has been made public, and instead of throwing an
exception, the StatusCode is set and read from the main program after
the Trainer exits its training function.