Implemented ZipParser, fully implemented RawParser
The RawParser has been implemented in Trainer.

ZipParser has been implemented, however the time it takes to load a
single image is ~10x longer than RawParser on average.  I think this is
because the locate_entry utility must scan through the zip file.  I'm
going to rework it using mz_zip instead of mz_zip_reader, as the former
allows for random access (though I'll also need to build a lookup table
for the images).